+++ pkg/client/time.go
@@ -1,34 +1,31 @@
-package jobsqueueapi
+package client
 
 import (
 	"strconv"
 	"time"
 )
 
-// TimeFormat used in Storage API.
-const TimeFormat = "2006-01-02T15:04:05-07:00"
-
-// Time is encoded/decoded in TimeFormat used in Storage API.
+// Time is encoded/decoded in TimeFormat with format RFC3339 used in Keboola APIs.
 type Time time.Time
 
 // UnmarshalJSON implements JSON decoding.
 func (t *Time) UnmarshalJSON(data []byte) (err error) {
-	now, err := time.ParseInLocation(`"`+TimeFormat+`"`, string(data), time.Local)
+	now, _ := time.ParseInLocation(`"`+time.RFC3339+`"`, string(data), time.Local)
 	*t = Time(now)
 	return
 }
 
 // MarshalJSON implements JSON encoding.
 func (t Time) MarshalJSON() ([]byte, error) {
-	b := make([]byte, 0, len(TimeFormat)+2)
+	b := make([]byte, 0, len(time.RFC3339)+2)
 	b = append(b, '"')
-	b = time.Time(t).AppendFormat(b, TimeFormat)
+	b = time.Time(t).AppendFormat(b, time.RFC3339)
 	b = append(b, '"')
 	return b, nil
 }
 
 func (t Time) String() string {
-	return time.Time(t).Format(TimeFormat)
+	return time.Time(t).Format(time.RFC3339)
 }
 
 // DurationSeconds is time.Duration encoded/decoded as number of seconds.
